src/requirements.txt
Flask==2.2.3
Pillow
requests
